+++ src/com/onsoft/tiger/jcad/decorators/AfterAllDecorator.ts
@@ -49,6 +49,6 @@ export class AfterAllDecorator implements Decorator {
     let methodDescriptor:AnnotatedMethodDescriptor =
           builder.build(key, descriptor, AnnotatedMethodType.AFTER_ALL, params);
     TestSuiteDescriptorRegistry.addAnnotatedMethodDescriptor(methodDescriptor);
-    return target;
+    return descriptor;
   }
 }